4088: Changed imports from AppContextProvider to AppContext across multiple components and reverted pushNotificationsManager
