+++ native/src/components/Header.tsx
@@ -24,7 +24,7 @@ import { config } from 'translations'
 import { ROOT_NAVIGATOR_ID, TAB_NAVIGATOR_ID } from '../constants'
 import { NavigationProps, RouteProps, RoutesParamsType, RoutesType } from '../constants/NavigationTypes'
 import dimensions from '../constants/dimensions'
-import { AppContext } from '../contexts/AppContextProvider'
+import { AppContext } from '../contexts/AppContext'
 import useSnackbar from '../hooks/useSnackbar'
 import useTtsPlayer from '../hooks/useTtsPlayer'
 import supportedLanguages from '../utils/supportedLanguages'
