+++ native/src/routes/Landing.tsx
@@ -12,7 +12,7 @@ import SwitchCmsUrlIcon from '../components/SwitchCmsUrlIcon'
 import Text from '../components/base/Text'
 import { NavigationProps } from '../constants/NavigationTypes'
 import buildConfig from '../constants/buildConfig'
-import { AppContext } from '../contexts/AppContextProvider'
+import { AppContext } from '../contexts/AppContext'
 import useLoadRegions from '../hooks/useLoadRegions'
 import testID from '../testing/testID'
 import dataContainer from '../utils/DefaultDataContainer'
