+++ native/src/utils/DatabaseConnector.ts
@@ -28,7 +28,7 @@ import {
   RESOURCE_CACHE_DIR_PATH,
   UNVERSIONED_CONTENT_DIR_PATH,
   UNVERSIONED_RESOURCE_CACHE_DIR_PATH,
-} from '../constants/DatabasePaths'
+} from '../constants/database'
 import DatabaseContext from '../models/DatabaseContext'
 import { RegionResourceCacheStateType } from './DataContainer'
 import { deleteIfExists } from './helpers'
