+++ native/src/utils/__tests__/PushNotificationsManager.spec.tsx
@@ -15,7 +15,6 @@ import TestingAppContext from '../../testing/TestingAppContext'
 import render from '../../testing/render'
 import * as PushNotificationsManager from '../PushNotificationsManager'
 import { usePushNotificationListener } from '../PushNotificationsManager'
-import { subscribeNews, unsubscribeNews } from '../PushNotificationsNews'
 
 jest.mock('@notifee/react-native', () => ({
   AndroidImportance: { HIGH: 4 },
@@ -52,7 +51,7 @@ describe('PushNotificationsManager', () => {
       const mockUnsubscribeFromTopic = jest.fn()
       mocked(unsubscribeFromTopic).mockImplementation(mockUnsubscribeFromTopic)
 
-      await unsubscribeNews('augsburg', 'de')
+      await PushNotificationsManager.unsubscribeNews('augsburg', 'de')
       expect(mockUnsubscribeFromTopic).toHaveBeenCalledWith('messaging', 'augsburg-de-news')
       expect(mockUnsubscribeFromTopic).toHaveBeenCalledTimes(1)
     })
@@ -63,7 +62,7 @@ describe('PushNotificationsManager', () => {
       const mockSubscribeToTopic = jest.fn()
       mocked(subscribeToTopic).mockImplementation(mockSubscribeToTopic)
 
-      await subscribeNews({
+      await PushNotificationsManager.subscribeNews({
         regionCode: 'augsburg',
         languageCode: 'de',
         allowPushNotifications: false,
@@ -75,7 +74,7 @@ describe('PushNotificationsManager', () => {
       const mockSubscribeToTopic = jest.fn()
       mocked(subscribeToTopic).mockImplementation(mockSubscribeToTopic)
 
-      await subscribeNews({
+      await PushNotificationsManager.subscribeNews({
         regionCode: 'augsburg',
         languageCode: 'de',
         allowPushNotifications: true,
@@ -88,7 +87,7 @@ describe('PushNotificationsManager', () => {
       const mockSubscribeToTopic = jest.fn()
       mocked(subscribeToTopic).mockImplementation(mockSubscribeToTopic)
 
-      await subscribeNews({
+      await PushNotificationsManager.subscribeNews({
         regionCode: 'augsburg',
         languageCode: 'de',
         allowPushNotifications: false,
