+++ native/src/utils/createSettingsSections.ts
@@ -10,10 +10,9 @@ import { SnackbarType } from '../components/SnackbarContainer'
 import NativeConstants from '../constants/NativeConstants'
 import { NavigationProps } from '../constants/NavigationTypes'
 import buildConfig from '../constants/buildConfig'
-import { AppContextType } from '../contexts/AppContextProvider'
+import { AppContextType } from '../contexts/AppContext'
 import { SettingsType } from './AppSettings'
-import { requestPushNotificationPermission } from './PushNotificationsManager'
-import { subscribeNews, unsubscribeNews } from './PushNotificationsNews'
+import { requestPushNotificationPermission, subscribeNews, unsubscribeNews } from './PushNotificationsManager'
 import openExternalUrl from './openExternalUrl'
 import { initSentry } from './sentry'
