+++ native/src/utils/helpers.ts
@@ -3,8 +3,8 @@ import { Platform } from 'react-native'
 import BlobUtil from 'react-native-blob-util'
 import Url from 'url-parse'
 
-import { RESOURCE_CACHE_DIR_PATH } from '../constants/DatabasePaths'
 import buildConfig from '../constants/buildConfig'
+import { RESOURCE_CACHE_DIR_PATH } from '../constants/database'
 import appSettings from './AppSettings'
 import { LanguageResourceCacheStateType } from './DataContainer'
 import { log } from './sentry'
